:zap: • Refactored and optimized the code
